refactor: move favourite and saved-search helpers into src/util
The web layer needs both, and eslint forbids src/web importing src/ui
(they are two front-ends over src/core). Same move sort.ts and filter.ts
made when the web search pane needed them.

Pure rename plus import updates; no assertion changed.
